🪛[refactor]: Management → Staff 네이밍 변경 및 파일명 컨벤션 정리
