+++ Projects/Shared/DesignSystem/Sources/Extension/Color/UIColor+.swift
@@ -1,14 +1,14 @@
 //
-//  Extension+Color.swift
+//  UIColor+.swift
 //  DDDAttendance
 //
 //  Created by 서원지 on 6/6/24.
 //
 
-import SwiftUI
+import UIKit
 
-public extension Color {
-  init(hex: String) {
+public extension UIColor {
+  convenience init(hex: String, alpha: Double? = .zero) {
     let scanner = Scanner(string: hex)
     _ = scanner.scanString("#")
     
@@ -18,6 +18,6 @@ public extension Color {
     let r = Double((rgb >> 16) & 0xFF) / 255.0
     let g = Double((rgb >>  8) & 0xFF) / 255.0
     let b = Double((rgb >>  0) & 0xFF) / 255.0
-    self.init(red: r, green: g, blue: b)
+    self.init(red: r, green: g, blue: b, alpha: 100)
   }
 }
